hashtags on edited text to lower case
